Add files via upload
This project is a **Cafeteria Management** that dynamically displays different menus (morning and evening) based on the time of day, adjusted for Indian Standard Time (IST). It allows users to place orders, calculates the total bill including GST, and generates a receipt with customer and payment details. The program also supports multiple orders, offering an efficient solution for coffee shop billing management.